+++ Code/code/docs/.vitepress/theme/index.ts
@@ -1,0 +1,12 @@
+import DefaultTheme from 'vitepress/theme'
+import RunGuide from './components/RunGuide.vue'
+import QuickJump from './components/QuickJump.vue'
+import './custom.css'
+
+export default {
+  extends: DefaultTheme,
+  enhanceApp({ app }) {
+    app.component('RunGuide', RunGuide)
+    app.component('QuickJump', QuickJump)
+  }
+}